Isolate service HttpException and implement core services
