+++ src/domain/categories/categories.module.ts
@@ -1,12 +1,14 @@
 import { Module } from '@nestjs/common';
-import { CategoriesService } from './categories.service';
+import { CategoriesService } from './core/categories.service';
 import { CategoriesController } from './controller/categories.controller';
 import { ICategoryRepository } from './repositories/ICategoryRepository';
-import { CategoryModelRepository } from '../../frameworks/database/postgres/categories/repositories/category.repository';
+import { CategoryModelRepository } from '../../infra/databases/postgres/categories/repositories/category.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CategoryModel } from '../../frameworks/database/postgres/categories/models/category.model';
-import { ICategoriesService } from './ports/ICategoriesService';
+import { CategoryModel } from '../../infra/databases/postgres/categories/models/category.model';
+import { ICategoriesService } from './core/categories.service.interface';
 import { CategoryMapper } from './mappers/category.mapper';
+import { ExceptionsService } from 'src/infra/exceptions/exceptions.service';
+import { IExceptionService } from '../shared/exceptions/exceptions.interface';
 
 @Module({
   imports: [TypeOrmModule.forFeature([CategoryModel])],
@@ -23,6 +25,10 @@ import { CategoryMapper } from './mappers/category.mapper';
     {
       provide: ICategoriesService,
       useClass: CategoriesService,
+    },
+    {
+      provide: IExceptionService,
+      useClass: ExceptionsService,
     },
     CategoryMapper,
   ],
