+++ src/domain/categories/controller/categories.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { CategoriesController } from './categories.controller';
-import { CategoriesService } from '../categories.service';
+import { CategoriesService } from '../core/categories.service';
 
 describe('CategoriesController', () => {
   let controller: CategoriesController;
