+++ src/domain/categories/core/categories.service.interface.ts
