+++ src/domain/customers/controller/customers.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { CustomersController } from './customers.controller';
-import { CustomersService } from '../customers.service';
+import { CustomersService } from '../core/customers.service';
 
 describe('CustomersController', () => {
   let controller: CustomersController;
