+++ src/domain/customers/customers.module.ts
@@ -1,11 +1,13 @@
 import { Module } from '@nestjs/common';
-import { CustomersService } from './customers.service';
+import { CustomersService } from './core/customers.service';
 import { CustomersController } from './controller/customers.controller';
 import { ICustomerRepository } from './repositories/ICustomerRepository';
-import { CustomerModelRepository } from '../../frameworks/database/postgres/customers/repositories/customer.repository';
+import { CustomerModelRepository } from '../../infra/databases/postgres/customers/repositories/customer.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CustomerModel } from '../../frameworks/database/postgres/customers/models/customer.model';
-import { ICustomersService } from './ports/ICustomersService';
+import { CustomerModel } from '../../infra/databases/postgres/customers/models/customer.model';
+import { ICustomersService } from './core/customers.service.interface';
+import { IExceptionService } from '../shared/exceptions/exceptions.interface';
+import { ExceptionsService } from 'src/infra/exceptions/exceptions.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([CustomerModel])],
@@ -22,6 +24,10 @@ import { ICustomersService } from './ports/ICustomersService';
     {
       provide: ICustomersService,
       useClass: CustomersService,
+    },
+    {
+      provide: IExceptionService,
+      useClass: ExceptionsService,
     },
   ],
 })
