+++ src/domain/orders/orders.module.ts
@@ -1,19 +1,19 @@
 import { Module } from '@nestjs/common';
-import { OrdersService } from './orders.service';
+import { OrdersService } from './core/orders.service';
 import { OrdersController } from './controller/orders.controller';
 import { IOrderRepository } from './repositories/IOrderRepository';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { OrderModel } from '../../frameworks/database/postgres/orders/models/order.model';
-import { IOrdersService } from './ports/IOrdersService';
-import { OrderModelRepository } from '../../frameworks/database/postgres/orders/repositories/orders.repository';
+import { OrderModel } from '../../infra/databases/postgres/orders/models/order.model';
+import { IOrdersService } from './core/orders.service.interface';
+import { OrderModelRepository } from '../../infra/databases/postgres/orders/repositories/orders.repository';
 import { ICustomerRepository } from '../customers/repositories/ICustomerRepository';
-import { CustomerModelRepository } from '../../frameworks/database/postgres/customers/repositories/customer.repository';
+import { CustomerModelRepository } from '../../infra/databases/postgres/customers/repositories/customer.repository';
 import { IProductRepository } from '../products/respositories/IProductRepository';
-import { ProductModelRepository } from '../../frameworks/database/postgres/products/repositories/product.repository';
-import { OrdersProductsAmountsModel } from '../../frameworks/database/postgres/orders/models/orders_products_amounts.model';
-import { CategoryModel } from '../../frameworks/database/postgres/categories/models/category.model';
-import { CustomerModel } from '../../frameworks/database/postgres/customers/models/customer.model';
-import { ProductModel } from '../../frameworks/database/postgres/products/models/product.model';
+import { ProductModelRepository } from '../../infra/databases/postgres/products/repositories/product.repository';
+import { OrdersProductsAmountsModel } from '../../infra/databases/postgres/orders/models/orders_products_amounts.model';
+import { CategoryModel } from '../../infra/databases/postgres/categories/models/category.model';
+import { CustomerModel } from '../../infra/databases/postgres/customers/models/customer.model';
+import { ProductModel } from '../../infra/databases/postgres/products/models/product.model';
 
 @Module({
   imports: [
