+++ src/domain/products/controllers/products.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ProductsController } from './products.controller';
-import { ProductsService } from '../products.service';
+import { ProductsService } from '../core/products.service';
 
 describe('ProductsController', () => {
   let controller: ProductsController;
