+++ src/domain/products/products.module.ts
@@ -1,16 +1,18 @@
 import { Module } from '@nestjs/common';
-import { ProductsService } from './products.service';
+import { ProductsService } from './core/products.service';
 import { ProductsController } from './controllers/products.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ProductModel } from '../../frameworks/database/postgres/products/models/product.model';
-import { ProductImageModel } from '../../frameworks/database/postgres/products/models/productImage.model';
+import { ProductModel } from '../../infra/databases/postgres/products/models/product.model';
+import { ProductImageModel } from '../../infra/databases/postgres/products/models/productImage.model';
 import { ICategoryRepository } from '../categories/repositories/ICategoryRepository';
-import { CategoryModelRepository } from '../../frameworks/database/postgres/categories/repositories/category.repository';
+import { CategoryModelRepository } from '../../infra/databases/postgres/categories/repositories/category.repository';
 import { IProductRepository } from './respositories/IProductRepository';
-import { ProductModelRepository } from '../../frameworks/database/postgres/products/repositories/product.repository';
-import { CategoryModel } from '../../frameworks/database/postgres/categories/models/category.model';
-import { IProductService } from './ports/IProductService';
+import { ProductModelRepository } from '../../infra/databases/postgres/products/repositories/product.repository';
+import { CategoryModel } from '../../infra/databases/postgres/categories/models/category.model';
+import { IProductService } from './core/products.service.interface';
 import { CategoryMapper } from '../categories/mappers/category.mapper';
+import { IExceptionService } from '../shared/exceptions/exceptions.interface';
+import { ExceptionsService } from 'src/infra/exceptions/exceptions.service';
 
 @Module({
   imports: [
@@ -29,6 +31,10 @@ import { CategoryMapper } from '../categories/mappers/category.mapper';
     {
       provide: IProductRepository,
       useClass: ProductModelRepository,
+    },
+    {
+      provide: IExceptionService,
+      useClass: ExceptionsService,
     },
     CategoryMapper,
   ],
