+++ src/domain/shared/exceptions/exceptions.interface.ts
@@ -1,0 +1,13 @@
+export interface IExceptionBody {
+  message: string;
+  code?: number;
+}
+
+export interface IExceptionService {
+  badRequestException(data: IExceptionBody): void;
+  forbiddenException(data?: IExceptionBody): void;
+  notFoundException(data?: IExceptionBody): void;
+  internalServerErrorException(data?: IExceptionBody): void;
+}
+
+export const IExceptionService = Symbol('IExceptionService');
